% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckARCoef}
\alias{CheckARCoef}
\title{Check AR(p) Coefficients for Stationarity}
\usage{
CheckARCoef(coef)
}
\arguments{
\item{coef}{Numeric vector.
Autoregressive coefficients.}
}
\description{
This function checks for stationarity of the AR(p) coefficients.
Stationarity is determined based on the roots
of the autoregressive polynomial.
For a stationary AR(p) process,
all the roots of this autoregressive polynomial
must lie inside the unit circle in the complex plane.
}
\examples{
set.seed(42)
(coef <- SimARCoef(p = 2))
CheckARCoef(coef = coef)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckVARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{ar}
\keyword{check}
\keyword{simAutoReg}
