% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CheckVARCoef}
\alias{CheckVARCoef}
\title{Check VAR(p) Coefficients for Stationarity}
\usage{
CheckVARCoef(coef)
}
\arguments{
\item{coef}{Numeric matrix.
Coefficient matrix with dimensions \code{k} by \code{(k * p)}.
Each \code{k} by \code{k} block corresponds to the coefficient matrix
for a particular lag.}
}
\description{
The function checks if all the eigenvalues have moduli
(absolute values) less than 1.
If all eigenvalues have moduli less than 1,
it indicates that the VAR process is stable and, therefore, stationary.
If any eigenvalue has a modulus greater than or equal to 1,
it indicates that the VAR process is not stable and,
therefore, non-stationary.
}
\examples{
set.seed(42)
CheckVARCoef(SimVARCoef(k = 3, p = 2))

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVAR2}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo2}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIP2}()},
\code{\link{SimVARZIPExo2}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
