% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LambdaSeq}
\alias{LambdaSeq}
\title{Function to generate the sequence of lambdas}
\usage{
LambdaSeq(Y, X, n_lambdas = 100L)
}
\arguments{
\item{Y}{Numeric matrix.
Matrix of dependent variables (Y).}

\item{X}{Numeric matrix.
Matrix of predictors (X).}

\item{n_lambdas}{Integer.
Number of lambdas to generate.}
}
\value{
Returns a vector of lambdas.
}
\description{
Function to generate the sequence of lambdas
}
\examples{
Ystd <- StdMat(vark3p2yx$Y)
Xstd <- StdMat(vark3p2yx$X[, -1])
LambdaSeq(Y = Ystd, X = Xstd, n_lambdas = 100)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fit}
\keyword{simAutoReg}
