% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OrigScale}
\alias{OrigScale}
\title{Return Standardized Estimates to the Original Scale}
\usage{
OrigScale(coef_std, sd_Y, sd_X, mean_Y, mean_X)
}
\arguments{
\item{coef_std}{Numeric matrix.
Standardized estimates of the autoregression and cross regression coefficients.}

\item{sd_Y}{Numeric vector.
Standard deviations of the \code{Y} matrix.}

\item{sd_X}{Numeric vector.
Standard deviations of the \code{X} matrix.}

\item{mean_Y}{Numeric vector.
Means of the \code{Y} matrix.}

\item{mean_X}{Numeric vector.
Means of the \code{X} matrix.}
}
\description{
Return Standardized Estimates to the Original Scale
}
\examples{
coef_std <- FitVAROLS(Y = StdMat(VAR_YX$Y), X = StdMat(VAR_YX$X[, -1]))
sd_Y <- sqrt(diag(var(VAR_YX$Y)))
sd_X <- sqrt(diag(var(VAR_YX$X[, -1])))
mean_Y <- colMeans(VAR_YX$Y)
mean_X <- colMeans(VAR_YX$X[, -1])
OrigScale(coef_std = coef_std, sd_Y = sd_Y, sd_X = sd_X, mean_Y = mean_Y, mean_X = mean_X)
FitVAROLS(Y = VAR_YX$Y, X = VAR_YX$X[, -1])

}
\author{
Ivan Jacob Agaloos Pesigan
}
