% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAutoReg-p-boot-ci.R
\name{PBootCI}
\alias{PBootCI}
\title{Parametric Bootstrap Confidence Intervals}
\usage{
PBootCI(x, alpha = 0.05)
}
\arguments{
\item{x}{Numeric matrix.
Output of \code{\link[=PBootVAROLS]{PBootVAROLS()}}.}

\item{alpha}{Numeric.
Significance level.}
}
\value{
A list with two elements, namely \code{ll} for the lower limit
and \code{ul} for the upper limit.
}
\description{
Parametric Bootstrap Confidence Intervals
}
\examples{
pb <- PBootVAROLS(data = vark3p2, p = 2, B = 100)
PBootCI(pb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootSE}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{pb}
\keyword{simAutoReg}
