% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAutoReg-p-boot-se.R
\name{PBootSE}
\alias{PBootSE}
\title{Parametric Bootstrap Standard Errors}
\usage{
PBootSE(x)
}
\arguments{
\item{x}{Numeric matrix.
Output of \code{\link[=PBootVAROLS]{PBootVAROLS()}}.}
}
\value{
A matrix of standard error.
}
\description{
Parametric Bootstrap Standard Errors
}
\examples{
pb <- PBootVAROLS(data = VAR, p = 2, B = 100)
PBootSE(pb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{pb}
\keyword{simAutoReg}
