% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PBootVAROLS}
\alias{PBootVAROLS}
\title{Parametric Bootstrap for the Vector Autoregressive Model
Using Ordinary Least Squares}
\usage{
PBootVAROLS(data, p, B = 1000L, burn_in = 200L)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{B}{Integer.
Number of bootstrap samples to generate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results
in the simulation step.}
}
\value{
List containing the estimates (\code{est})
and bootstrap estimates (\code{boot}).
}
\description{
Parametric Bootstrap for the Vector Autoregressive Model
Using Ordinary Least Squares
}
\examples{
pb <- PBootVAROLS(data = vark3p2, p = 2, B = 10)
str(pb)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVARLasso}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{pb}
\keyword{simAutoReg}
