% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SearchVARLasso}
\alias{SearchVARLasso}
\title{Compute AIC, BIC, and EBIC for Lasso Regularization}
\usage{
SearchVARLasso(Ystd, Xstd, lambdas, max_iter = 10000L, tol = 1e-05)
}
\arguments{
\item{Ystd}{Numeric matrix.
Matrix of standardized dependent variables (Y).}

\item{Xstd}{Numeric matrix.
Matrix of standardized predictors (X).}

\item{lambdas}{Numeric vector.
Vector of lambda hyperparameters for Lasso regularization.}

\item{max_iter}{Integer.
The maximum number of iterations for the coordinate descent algorithm.
Default is 10000.}

\item{tol}{Numeric.
Convergence tolerance. The algorithm stops when the change in coefficients
between iterations is below this tolerance. Default is 1e-5.}
}
\value{
List containing two elements:
\itemize{
\item Element 1: Matrix with columns for
lambda, AIC, BIC, and EBIC values.
\item Element 2: List of matrices containing
the estimated autoregressive
and cross-regression coefficients for each lambda.
}
}
\description{
This function computes the Akaike Information Criterion (AIC),
Bayesian Information Criterion (BIC),
and Extended Bayesian Information Criterion (EBIC)
for a given matrix of predictors \code{X}, a matrix of outcomes \code{Y},
and a vector of lambda hyperparameters for Lasso regularization.
}
\examples{
Ystd <- StdMat(vark3p2yx$Y)
Xstd <- StdMat(vark3p2yx$X[, -1])
lambdas <- 10^seq(-5, 5, length.out = 100)
search <- SearchVARLasso(Ystd = Ystd, Xstd = Xstd, lambdas = lambdas)
plot(x = 1:nrow(search$criteria), y = search$criteria[, 4],
  type = "b", xlab = "lambda", ylab = "EBIC")

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fit}
\keyword{simAutoReg}
