% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAutoReg-select-var-lasso.R
\name{SelectVARLasso}
\alias{SelectVARLasso}
\title{Select the Lasso Estimates from the Grid Search}
\usage{
SelectVARLasso(search, crit = "ebic")
}
\arguments{
\item{search}{Object.
Output of the \code{\link[=SearchVARLasso]{SearchVARLasso()}} function.}

\item{crit}{Character string.
Information criteria to use.
Valid values include \code{"aic"}, \code{"bic"}, and \code{"ebic"}.}
}
\value{
Returns the Lasso estimates
of autoregression and cross regression coefficients.
}
\description{
Select the Lasso Estimates from the Grid Search
}
\examples{
Y_std <- StdMat(VAR_YX$Y)
X_std <- StdMat(VAR_YX$X[, -1])
lambdas <- 10^seq(-5, 5, length.out = 100)
search <- SearchVARLasso(Y_std = Y_std, X_std = X_std, lambdas = lambdas)
SelectVARLasso(search, crit = "ebic")

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fit}
\keyword{simAutoReg}
