% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimAR}
\alias{SimAR}
\title{Simulate Data from an Autoregressive Model with Constant Term}
\usage{
SimAR(time, burn_in, constant, coef, sd)
}
\arguments{
\item{time}{Integer.
Number of time points to simulate.}

\item{burn_in}{Integer.
Number of burn-in periods before recording data.}

\item{constant}{Numeric.
The constant term of the AR model.}

\item{coef}{Numeric vector.
Autoregressive coefficients.}

\item{sd}{Numeric.
The standard deviation of the random process noise.}
}
\value{
Numeric vector (column matrix) containing the simulated
time series data.
}
\description{
This function generates synthetic time series data
from an autoregressive (AR) model.
}
\details{
The \code{\link[=SimAR]{SimAR()}} function generates synthetic time series data
from an autoregressive (AR) model.
The generated data follows the AR(p) model,
where \code{p} is the number of coefficients specified in \code{coef}.
The generated time series data includes a constant term
and autoregressive terms based on the provided coefficients.
Random noise, sampled from a normal distribution with mean 0
and standard deviation \code{sd}, is added to the time series.
A burn-in period is specified to exclude initial data points
from the output.

The steps in generating the autoregressive time series with burn-in
are as follows:
\itemize{
\item Set the order of the AR model to \code{p} based on the length of \code{coef}.
\item Create a vector data of size \code{time + burn_in}
to store the generated AR time series data.
\item Create a vector data of size \code{time + burn_in} of random process noise
from a normal distribution with mean 0
and standard deviation \code{sd}.
\item Generate the autoregressive time series with burn-in using the formula:
\deqn{
    Y_t = \mathrm{constant} +
    \sum_{i = 1}^{p} \left( \mathrm{coef}_i * Y_{t - i} \right) +
    \mathrm{noise}_t
  }
where \eqn{Y_t} is the time series data at time \eqn{t},
\eqn{\mathrm{constant}} is the constant term,
\eqn{\mathrm{coef}_i} are the autoregressive coefficients,
\eqn{Y_{t - i}} are the lagged data points up to order \code{p},
and \eqn{\mathrm{noise}_t} is the random noise at time \eqn{t}.
\item Remove the burn-in period from the generated time series data.
}
}
\examples{
set.seed(42)
SimAR(time = 10, burn_in = 5, constant = 2, coef = c(0.5, -0.3), sd = 0.1)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{CheckVARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
