% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimARCoef}
\alias{SimARCoef}
\title{Simulate Random Autoregressive Coefficients for a Stationary AR(p) Model}
\usage{
SimARCoef(p)
}
\arguments{
\item{p}{Positive integer. Number of lags.}
}
\description{
This function generates autoregressive coefficients
for a stationary AR(p) model.
}
\examples{
set.seed(42)
SimARCoef(p = 2)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{CheckVARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
