% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAutoReg-package.R
\docType{package}
\name{simAutoReg-package}
\alias{simAutoReg-package}
\alias{_PACKAGE}
\title{simAutoReg: Simulate Data from Autoregressive Models}
\description{
Simulate data from autoregressive models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ijapesigan/simAutoReg}
  \item \url{https://ijapesigan.github.io/simAutoReg/}
  \item Report bugs at \url{https://github.com/ijapesigan/simAutoReg/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{ijapesigan@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
