% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimMVN}
\alias{SimMVN}
\title{Simulate Multivariate Normal Random Numbers}
\usage{
SimMVN(n, location, chol_scale)
}
\arguments{
\item{n}{Integer.
Number of samples to generate.}

\item{location}{Numeric vector.
Mean vector of length \code{k}, where \code{k} is the number of variables.}

\item{chol_scale}{Numeric matrix.
Cholesky decomposition of the covariance matrix of dimensions \code{k} by \code{k}.}
}
\value{
Matrix containing the simulated multivariate normal random numbers,
with dimensions \code{n} by \code{k}, where \code{n} is the number of samples
and \code{k} is the number of variables.
}
\description{
This function generates multivariate normal random numbers.
}
\details{
The \code{\link[=SimMVN]{SimMVN()}} function generates
multivariate normal random numbers
using the Cholesky decomposition method.
Given the number of samples \code{n}, the mean vector \code{location} of length \code{k}
(where \code{k} is the number of variables),
and the Cholesky decomposition \code{chol_scale} of the covariance matrix
of dimensions \code{k} by \code{k},
the function produces a matrix of multivariate normal random numbers.

The steps involved in generating multivariate normal random numbers
are as follows:
\itemize{
\item Determine the number of variables \code{k} from the length of the mean vector.
\item Generate random data from a standard multivariate normal distribution,
resulting in an \code{n} by \code{k} matrix of random numbers.
\item Transform the standard normal random data
into multivariate normal random data
using the Cholesky decomposition \code{chol_scale}.
\item Add the mean vector \code{location} to the transformed data
to obtain the final simulated multivariate normal random numbers.
\item The function returns a matrix of simulated
multivariate normal random numbers
with dimensions \code{n} by \code{k},
where \code{n} is the number of samples and \code{k} is the number of variables.
This matrix can be used for various statistical analyses and simulations.
}
}
\examples{
set.seed(42)
n <- 1000L
location <- c(0.5, -0.2, 0.1)
scale <- matrix(
  data = c(1.0, 0.3, 0.3, 0.3, 1.0, 0.2, 0.3, 0.2, 1.0),
  nrow = 3,
  byrow = TRUE
)
chol_scale <- chol(scale)
y <- SimMVN(n = n, location = location, chol_scale = chol_scale)
colMeans(y)
var(y)

}
\seealso{
The \code{\link[=chol]{chol()}} function in R to obtain the Cholesky decomposition
of a covariance matrix.

Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{CheckVARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimPD}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{data}
\keyword{mvn}
\keyword{sim}
\keyword{simAutoReg}
