% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimMVNFixed}
\alias{SimMVNFixed}
\title{Simulate Multivariate Normal Random Numbers with Optional Fixed Values}
\usage{
SimMVNFixed(n, location, scale)
}
\arguments{
\item{n}{Integer.
Number of samples to generate.}

\item{location}{Numeric vector.
Mean vector of length \code{k}, where \code{k} is the number of variables.}

\item{scale}{Numeric matrix.
Covariance matrix of dimensions \code{k} by \code{k}.
Values for variables with variance of \code{0} will be fixed
to the corresponding value in \code{location}.}
}
\value{
Matrix containing the simulated multivariate normal random numbers,
with dimensions \code{n} by \code{k}, where \code{n} is the number of samples
and \code{k} is the number of variables.
}
\description{
This function generates multivariate normal random numbers
with optinal fixed values where the variance is zero.
}
\details{
The \code{\link[=SimMVNFixed]{SimMVNFixed()}} function first identifies the indices
of non-constant variables (i.e., variables with variance not equal to 0)
in the covariance matrix.
It then extracts the non-constant elements from the mean vector
and the covariance matrix.
A Cholesky decomposition is performed on the covariance matrix
of non-constant variables.
Random samples are generated for the non-constant variables
using the Cholesky factor.
The generated data matrix is constructed
by setting the non-constant variables
and constant variables to their corresponding values.
}
\examples{
set.seed(42)
n <- 1000L
location <- c(0.5, -0.2, 0.1)
scale <- matrix(
  data = c(0, 0.3, 0.3, 0.3, 1.0, 0.2, 0.3, 0.2, 1.0),
  nrow = 3,
  byrow = TRUE
)
y <- SimMVNFixed(n = n, location = location, scale = scale)
colMeans(y)
var(y)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
