% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimPD}
\alias{SimPD}
\title{Simulate Random Symmetric Positive Definite Matrix}
\usage{
SimPD(p)
}
\arguments{
\item{p}{Positive integer.
Dimension of the \code{p} by \code{p} matrix.}
}
\description{
This function generates a random positive definite matrix.
}
\examples{
set.seed(42)
SimPD(p = 3)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
