% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimVARCoef}
\alias{SimVARCoef}
\title{Simulate Random Vector Autoregressive Coefficients
for a Stationary VAR(p) Model}
\usage{
SimVARCoef(k, p)
}
\arguments{
\item{k}{Positive integer. Number of autoregressive variables.}

\item{p}{Positive integer. Number of lags.}
}
\description{
This function generates stationary VAR(P) coefficients.
}
\examples{
set.seed(42)
SimVARCoef(k = 3, p = 2)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{CheckVARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
