% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimVARExo}
\alias{SimVARExo}
\title{Simulate Data from a Vector Autoregressive (VAR) Model with Exogenous
Variables}
\usage{
SimVARExo(time, burn_in, constant, coef, chol_cov, exo_mat, exo_coef)
}
\arguments{
\item{time}{Integer.
Number of time points to simulate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results.}

\item{constant}{Numeric vector.
The constant term vector of length \code{k},
where \code{k} is the number of variables.}

\item{coef}{Numeric matrix.
Coefficient matrix with dimensions \code{k} by \code{(k * p)}.
Each \code{k} by \code{k} block corresponds to the coefficient matrix
for a particular lag.}

\item{chol_cov}{Numeric matrix.
The Cholesky decomposition of the covariance matrix
of the multivariate normal noise.
It should have dimensions \code{k} by \code{k}.}

\item{exo_mat}{Numeric matrix.
Matrix of exogenous covariates with dimensions \code{time + burn_in} by \code{x}.
Each column corresponds to a different exogenous variable.}

\item{exo_coef}{Numeric vector.
Coefficient matrix with dimensions \code{k} by \code{x}
associated with the exogenous covariates.}
}
\value{
Numeric matrix containing the simulated time series data
with dimensions \code{k} by \code{time},
where \code{k} is the number of variables and
\code{time} is the number of observations.
}
\description{
This function generates synthetic time series data
from a Vector Autoregressive (VAR) model with exogenous variables.
}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{CheckVARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVAR2}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo2}()},
\code{\link{SimVARZIP2}()},
\code{\link{SimVARZIPExo2}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
