% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimVariance}
\alias{SimVariance}
\title{Generate Random Data for the Variance Vector}
\usage{
SimVariance(n, location, chol_scale)
}
\arguments{
\item{n}{Integer.
Number of samples to generate.}

\item{location}{Numeric vector.
The constant term \eqn{\boldsymbol{\mu}}.}

\item{chol_scale}{Numeric matrix.
Cholesky decomposition of the covariance matrix \eqn{\boldsymbol{\Sigma}}
for the multivariate normal random error \eqn{\boldsymbol{\varepsilon}}.}
}
\value{
Matrix with each row containing the simulated variance vector
for each sample.
}
\description{
This function generates random data for the variance vector given by
\deqn{
  \boldsymbol{\sigma^{2}} =
  \exp \left( \boldsymbol{\mu} + \boldsymbol{\varepsilon} \right)
  \quad
  \text{with}
  \boldsymbol{\varepsilon} \sim
  \mathcal{N} \left( \boldsymbol{0}, \boldsymbol{\Sigma} \right)
}.
}
\examples{
set.seed(42)
n <- 100
location <- c(0.5, -0.2, 0.1)
chol_scale <- chol(
  matrix(
    data = c(1.0, 0.3, 0.3, 0.3, 1.0, 0.2, 0.3, 0.2, 1.0),
    nrow = 3,
    byrow = TRUE
  )
)
SimVariance(n = n, location = location, chol_scale = chol_scale)

}
\author{
Ivan Jacob Agaloos Pesigan
}
