% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimVARZIP}
\alias{SimVARZIP}
\title{Simulate Data from a Vector Autoregressive Zero-Inflated Poisson (VARZIP)
Model}
\usage{
SimVARZIP(time, burn_in, constant, coef, chol_cov)
}
\arguments{
\item{time}{Integer.
Number of time points to simulate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results.}

\item{constant}{Numeric vector.
The constant term vector of length \code{k},
where \code{k} is the number of variables.}

\item{coef}{Numeric matrix.
Coefficient matrix with dimensions \code{k} by \code{(k * p)}.
Each \code{k} by \code{k} block corresponds to the coefficient matrix
for a particular lag.}

\item{chol_cov}{Numeric matrix.
The Cholesky decomposition of the covariance matrix
of the multivariate normal noise.
It should have dimensions \code{k} by \code{k}.}
}
\value{
Numeric matrix containing the simulated time series data
with dimensions \code{k} by \code{(time - burn_in)},
where \code{k} is the number of variables
and time is the number of observations.
}
\description{
This function generates synthetic time series data
from a Vector Autoregressive Zero-Inflated Poisson (VARZIP) model.
}
\details{
The \code{\link[=SimVARZIP]{SimVARZIP()}} function generates synthetic time series data
from a Vector Autoregressive (VAR)
with Zero-Inflated Poisson (ZIP) model for the first observed variable.
See \code{\link[=SimVAR]{SimVAR()}} for more details on generating data for VAR(p).
The \code{SimVARZIP} function goes further by using the generated values
for the first variable to generate data from the ZIP model.
The exponential of the values from the first variable
from the original VAR(p) model
are used as the \code{intensity} parameter in the Poisson distribution
in the ZIP model.
Data from the ZIP model are used to replace the original values
for the first variable.
Values for the rest of the variables are unchanged.
The generated data includes a burn-in period,
which is excluded before returning the results.

The steps involved in generating the time series data are as follows:
\itemize{
\item Extract the number of variables \code{k}
and the number of lags \code{p} from the input.
\item Create a matrix \code{data} of size \code{k} x (\code{time + burn_in})
to store the generated data.
\item Set the initial values of the matrix \code{data}
using the constant term \code{constant}.
\item For each time point starting from the \code{p}-th time point
to \code{time + burn_in - 1}:
\itemize{
\item Generate a vector of random process noise
from a multivariate normal distribution
with mean 0 and covariance matrix \code{chol_cov}.
\item Generate the VAR time series values for each variable \code{j}
at time \code{t} by applying the autoregressive terms
for each lag \code{lag} and each variable \code{l}.
\itemize{
\item Add the generated noise to the VAR time series values.
\item For the first variable,
apply the Zero-Inflated Poisson (ZIP) model:
\itemize{
\item Compute the intensity \code{intensity}
as the exponential of the first variable's value at time \code{t}.
\item Sample a random value \code{u}
from a uniform distribution on [0, 1].
\item If \code{u} is less than \code{intensity / (1 + intensity)},
set the first variable's value to zero (inflation).
\item Otherwise, sample the first variable's value
from a Poisson distribution
with mean \code{intensity} (count process).
}
}
}
\item Transpose the data matrix \code{data} and return only
the required time period after burn-in as a numeric matrix.
}
}
\examples{
set.seed(42)
time <- 50L
burn_in <- 10L
k <- 3
p <- 2
constant <- c(1, 1, 1)
coef <- matrix(
  data = c(
    0.4, 0.0, 0.0, 0.1, 0.0, 0.0,
    0.0, 0.5, 0.0, 0.0, 0.2, 0.0,
    0.0, 0.0, 0.6, 0.0, 0.0, 0.3
  ),
  nrow = k,
  byrow = TRUE
)
chol_cov <- chol(diag(3))
y <- SimVARZIP(
  time = time,
  burn_in = burn_in,
  constant = constant,
  coef = coef,
  chol_cov = chol_cov
)
head(y)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
