% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAutoReg-vark3p2yx.R
\docType{data}
\name{vark3p2yx}
\alias{vark3p2yx}
\title{Data from the Vector Autoregressive Model (Y) and Lagged Predictors (X)}
\format{
A list with elements Y and X where Y is equal to the \code{vark3p2} data set
minus p = 2 terminal rows
and \code{X} is a matrix of ones for the first column and lagged values of \code{Y}
for the rest of the columns.
}
\usage{
vark3p2yx
}
\description{
Data from the Vector Autoregressive Model (Y) and Lagged Predictors (X)
}
\keyword{data}
\keyword{simAutoReg}
