% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitVARLassoSearch}
\alias{FitVARLassoSearch}
\title{Fit Vector Autoregressive (VAR) Model Parameters using Lasso Regularization
with Lambda Search}
\usage{
FitVARLassoSearch(
  Ystd,
  Xstd,
  lambdas,
  crit = "ebic",
  max_iter = 10000L,
  tol = 1e-05
)
}
\arguments{
\item{Ystd}{Numeric matrix.
Matrix of standardized dependent variables (Y).}

\item{Xstd}{Numeric matrix.
Matrix of standardized predictors (X).}

\item{lambdas}{Numeric vector.
Vector of lambda hyperparameters for Lasso regularization.}

\item{crit}{Character string.
Information criteria to use.
Valid values include \code{"aic"}, \code{"bic"}, and \code{"ebic"}.}

\item{max_iter}{Integer.
The maximum number of iterations for the coordinate descent algorithm.
Default is 10000.}

\item{tol}{Numeric.
Convergence tolerance. The algorithm stops when the change in coefficients
between iterations is below this tolerance. Default is 1e-5.}
}
\value{
Matrix of estimated autoregressive
and cross-regression coefficients.
}
\description{
Fit Vector Autoregressive (VAR) Model Parameters using Lasso Regularization
with Lambda Search
}
\examples{
Ystd <- StdMat(vark3p2yx$Y)
Xstd <- StdMat(vark3p2yx$X[, -1])
lambdas <- LambdaSeq(Y = Ystd, X = Xstd, n_lambdas = 100)
FitVARLassoSearch(Ystd = Ystd, Xstd = Xstd, lambdas = lambdas)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fit}
\keyword{simAutoReg}
