% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FitVAROLS}
\alias{FitVAROLS}
\title{Fit Vector Autoregressive (VAR) Model Parameters using OLS}
\usage{
FitVAROLS(Y, X)
}
\arguments{
\item{Y}{Numeric matrix.
Matrix of dependent variables (Y).}

\item{X}{Numeric matrix.
Matrix of predictors (X).}
}
\value{
Matrix of estimated autoregressive and cross-regression coefficients.
}
\description{
This function estimates the parameters of a VAR model
using the Ordinary Least Squares (OLS) method.
The OLS method is used to estimate the autoregressive
and cross-regression coefficients.
}
\details{
The \code{\link[=FitVAROLS]{FitVAROLS()}} function estimates the parameters
of a Vector Autoregressive (VAR) model
using the Ordinary Least Squares (OLS) method.
Given the input matrices \code{Y} and \code{X},
where \code{Y} is the matrix of dependent variables,
and \code{X} is the matrix of predictors,
the function computes the autoregressive and cross-regression coefficients
of the VAR model.
Note that if the first column of \code{X} is a vector of ones,
the constant vector is also estimated.

The steps involved in estimating the VAR model parameters using OLS are as follows:
\itemize{
\item Compute the QR decomposition of the lagged predictor matrix \code{X}
using the \code{qr} function from the Armadillo library.
\item Extract the \code{Q} and \code{R} matrices from the QR decomposition.
\item Solve the linear system \code{R * coef = Q.t() * Y}
to estimate the VAR model coefficients \code{coef}.
\item The function returns a matrix containing the estimated
autoregressive and cross-regression coefficients of the VAR model.
}
}
\examples{
FitVAROLS(Y = VAR_YX$Y, X = VAR_YX$X)

}
\seealso{
The \code{qr} function from the Armadillo library for QR decomposition.

Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLasso}()},
\code{\link{OrigScale}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{fit}
\keyword{simAutoReg}
