% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OrigScale}
\alias{OrigScale}
\title{Return Standardized Estimates to the Original Scale}
\usage{
OrigScale(coef_std, Y, X)
}
\arguments{
\item{coef_std}{Numeric matrix.
Standardized estimates of the autoregression
and cross regression coefficients.}

\item{Y}{Numeric matrix.
Matrix of dependent variables (Y).}

\item{X}{Numeric matrix.
Matrix of predictors (X).}
}
\description{
Return Standardized Estimates to the Original Scale
}
\examples{
coef_std <- FitVAROLS(Y = StdMat(VAR_YX$Y), X = StdMat(VAR_YX$X[, -1]))
OrigScale(coef_std = coef_std, Y = VAR_YX$Y, X = VAR_YX$X[, -1])
FitVAROLS(Y = VAR_YX$Y, X = VAR_YX$X[, -1])

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{simAutoReg}
\keyword{utils}
