% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimVAR}
\alias{SimVAR}
\title{Simulate Data from a Vector Autoregressive (VAR) Model}
\usage{
SimVAR(time, burn_in, constant, coef, chol_cov)
}
\arguments{
\item{time}{Integer.
Number of time points to simulate.}

\item{burn_in}{Integer.
Number of burn-in observations to exclude before returning the results.}

\item{constant}{Numeric vector.
The constant term vector of length \code{k},
where \code{k} is the number of variables.}

\item{coef}{Numeric matrix.
Coefficient matrix with dimensions \code{k} by \code{(k * p)}.
Each \code{k} by \code{k} block corresponds to the coefficient matrix
for a particular lag.}

\item{chol_cov}{Numeric matrix.
The Cholesky decomposition of the covariance matrix
of the multivariate normal noise.
It should have dimensions \code{k} by \code{k}.}
}
\value{
Numeric matrix containing the simulated time series data
with dimensions \code{k} by \code{time},
where \code{k} is the number of variables and
\code{time} is the number of observations.
}
\description{
This function generates synthetic time series data
from a Vector Autoregressive (VAR) model.
}
\details{
The \code{\link[=SimVAR]{SimVAR()}} function generates synthetic time series data
from a Vector Autoregressive (VAR) model.
The VAR model is defined by the constant term \code{constant},
the coefficient matrix \code{coef},
and the Cholesky decomposition of the covariance matrix
of the multivariate normal process noise \code{chol_cov}.
The generated time series data follows a VAR(p) process,
where \code{p} is the number of lags specified by the size of \code{coef}.
The generated data includes a burn-in period,
which is excluded before returning the results.

The steps involved in generating the VAR time series data are as follows:
\itemize{
\item Extract the number of variables \code{k} and the number of lags \code{p}
from the input.
\item Create a matrix \code{data} of size \code{k} by (\code{time + burn_in})
to store the generated VAR time series data.
\item Set the initial values of the matrix \code{data}
using the constant term \code{constant}.
\item For each time point starting from the \code{p}-th time point
to \code{time + burn_in - 1}:
\itemize{
\item Generate a vector of random noise
from a multivariate normal distribution
with mean 0 and covariance matrix \code{chol_cov}.
\item Generate the VAR time series values for each variable \code{j} at time \code{t}
using the formula:
\deqn{
      Y_{tj} = \mathrm{constant}_j +
      \sum_{l = 1}^{p} \sum_{m = 1}^{k} (\mathrm{coef}_{jm} * Y_{im}) +
      \mathrm{noise}_{j}
    }
where \eqn{Y_{tj}} is the value of variable \code{j} at time \code{t},
\eqn{\mathrm{constant}_j} is the constant term for variable \code{j},
\eqn{\mathrm{coef}_{jm}} are the coefficients for variable \code{j}
from lagged variables up to order \code{p},
\eqn{Y_{tm}} are the lagged values of variable \code{m}
up to order \code{p} at time \code{t},
and \eqn{\mathrm{noise}_{j}} is the element \code{j}
from the generated vector of random process noise.
}
\item Transpose the matrix \code{data} and return only
the required time period after the burn-in period,
which is from column \code{burn_in} to column \code{time + burn_in - 1}.
}
}
\examples{
set.seed(42)
time <- 50L
burn_in <- 10L
k <- 3
p <- 2
constant <- c(1, 1, 1)
coef <- matrix(
  data = c(
    0.4, 0.0, 0.0, 0.1, 0.0, 0.0,
    0.0, 0.5, 0.0, 0.0, 0.2, 0.0,
    0.0, 0.0, 0.6, 0.0, 0.0, 0.3
  ),
  nrow = k,
  byrow = TRUE
)
chol_cov <- chol(diag(3))
y <- SimVAR(
  time = time,
  burn_in = burn_in,
  constant = constant,
  coef = coef,
  chol_cov = chol_cov
)
head(y)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{CheckARCoef}()},
\code{\link{CheckVARCoef}()},
\code{\link{SimARCoef}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimPD}()},
\code{\link{SimVARCoef}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVariance}()},
\code{\link{YXExo}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{data}
\keyword{sim}
\keyword{simAutoReg}
\keyword{var}
