% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SimVariance}
\alias{SimVariance}
\title{Generate Random Data for the Variance Vector}
\usage{
SimVariance(n, location, chol_scale)
}
\arguments{
\item{n}{Integer.
Number of samples to generate.}

\item{location}{Numeric vector.
The constant term \eqn{\boldsymbol{\mu}}.}

\item{chol_scale}{Numeric matrix.
Cholesky decomposition of the covariance matrix \eqn{\boldsymbol{\Sigma}}
for the multivariate normal random error \eqn{\boldsymbol{\varepsilon}}.}
}
\value{
Matrix with each row containing the simulated variance vector
for each sample.
}
\description{
This function generates random data for the variance vector given by
\deqn{
  \boldsymbol{\sigma^{2}} =
  \exp \left( \boldsymbol{\mu} + \boldsymbol{\varepsilon} \right)
  \quad
  \text{with}
  \quad
  \boldsymbol{\varepsilon} \sim
  \mathcal{N} \left( \boldsymbol{0}, \boldsymbol{\Sigma} \right)
}.
}
\details{
The \code{\link[=SimVariance]{SimVariance()}} function generates random data
for the variance vector
based on the exponential of a multivariate normal distribution.
Given the number of samples \code{n},
the constant term \eqn{\boldsymbol{\mu}} represented
by the \code{location} vector,
and the Cholesky decomposition matrix \eqn{\boldsymbol{\Sigma}}
for the multivariate normal random error \eqn{\boldsymbol{\varepsilon}},
the function simulates \eqn{n} independent samples
of the variance vector \eqn{\boldsymbol{\sigma^{2}}}.
Each sample of the variance vector \eqn{\boldsymbol{\sigma^{2}}}
is obtained by
calculating the exponential of random variations
to the mean vector \eqn{\boldsymbol{\mu}}.
The random variations are generated using the Cholesky decomposition
of the covariance matrix \eqn{\boldsymbol{\Sigma}}.
Finally, the function returns a matrix with each column
containing the simulated
variance vector for each sample.
}
\examples{
set.seed(42)
n <- 100
location <- c(0.5, -0.2, 0.1)
chol_scale <- chol(
  matrix(
    data = c(1.0, 0.3, 0.3, 0.3, 1.0, 0.2, 0.3, 0.2, 1.0),
    nrow = 3,
    byrow = TRUE
  )
)
SimVariance(n = n, location = location, chol_scale = chol_scale)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVARLasso}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{StdMat}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{sim}
\keyword{simAutoReg}
