% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StdMat}
\alias{StdMat}
\title{Standardize Matrix}
\usage{
StdMat(X)
}
\arguments{
\item{X}{Numeric matrix.
The matrix to be standardized.}
}
\value{
Numeric matrix with standardized values.
}
\description{
This function standardizes the given matrix by centering the columns
and scaling them to have unit variance.
}
\examples{
std <- StdMat(VAR)
colMeans(std)
var(std)

}
\author{
Ivan Jacob Agaloos Pesigan
}
