% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{StdMat}
\alias{StdMat}
\title{Standardize Matrix}
\usage{
StdMat(X)
}
\arguments{
\item{X}{Numeric matrix.
The matrix to be standardized.}
}
\value{
Numeric matrix with standardized values.
}
\description{
This function standardizes the given matrix by centering the columns
and scaling them to have unit variance.
}
\examples{
std <- StdMat(vark3p2)
colMeans(std)
var(std)

}
\seealso{
Other Simulation of Autoregressive Data Functions: 
\code{\link{FitVARLassoSearch}()},
\code{\link{FitVARLasso}()},
\code{\link{FitVAROLS}()},
\code{\link{LambdaSeq}()},
\code{\link{OrigScale}()},
\code{\link{PBootCI}()},
\code{\link{PBootSE}()},
\code{\link{PBootVAROLS}()},
\code{\link{SearchVARLasso}()},
\code{\link{SelectVARLasso}()},
\code{\link{SimAR}()},
\code{\link{SimMVN}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{simAutoReg}
\keyword{utils}
