% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{YX}
\alias{YX}
\title{Create Y and X Matrices}
\usage{
YX(data, p)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions n x k,
where n is the number of observations and k is the number of variables.}

\item{p}{Integer.
The order of the VAR model (number of lags).}
}
\value{
List containing the Y and X matrices.
}
\description{
This function creates the Y and X matrices.
}
\details{
The \code{YX} function creates the Y and X matrices required for fitting a Vector Autoregressive (VAR) model.
Given the input \code{data} matrix with dimensions n x k, where n is the number of observations and k is the number of variables,
and the order of the VAR model \code{p} (number of lags), the function constructs lagged predictor matrix X and the dependent variable matrix Y.
The matrices X and Y are used as inputs for estimating the VAR model parameters.

The steps involved in creating the Y and X matrices are as follows:

\itemize{
\item Determine the number of observations \code{n} and the number of variables \code{k} from the input data matrix.
\item Create matrices X and Y to store lagged variables and the dependent variable, respectively.
\item Populate the matrices X and Y with the appropriate lagged data. The predictors matrix X contains the lagged values of the dependent variables,
while the dependent variable matrix Y contains the original values of the dependent variables.
}

The function returns a list containing the Y and X matrices, which can be used for further analysis and estimation of the VAR model parameters.
}
\examples{
set.seed(42)
time <- 100000L
burn_in <- 200
k <- 3
p <- 2
constant <- c(1, 1, 1)
coef <- matrix(
  data = c(
    0.4, 0.0, 0.0, 0.1, 0.0, 0.0,
    0.0, 0.5, 0.0, 0.0, 0.2, 0.0,
    0.0, 0.0, 0.6, 0.0, 0.0, 0.3
  ),
  nrow = k,
  byrow = TRUE
)
chol_cov <- chol(
  matrix(
    data = c(
      0.1, 0.0, 0.0,
      0.0, 0.1, 0.0,
      0.0, 0.0, 0.1
    ),
    nrow = k,
    byrow = TRUE
  )
)
y <- SimVAR(
  time = time,
  burn_in = burn_in,
  constant = constant,
  coef = coef,
  chol_cov = chol_cov
)
yx <- YX(data = y, p = p)
str(yx)

}
\seealso{
The \code{SimVAR} function for simulating time series data from a VAR model.
}
\author{
Ivan Jacob Agaloos Pesigan
}
