% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{YXExo}
\alias{YXExo}
\title{Create Y and X Matrices with Exogenous Variables}
\usage{
YXExo(data, p, exo_mat)
}
\arguments{
\item{data}{Numeric matrix.
The time series data with dimensions \code{t} by \code{k},
where \code{t} is the number of observations
and \code{k} is the number of variables.}

\item{p}{Integer.
The order of the VAR model (number of lags).}

\item{exo_mat}{Numeric matrix.
Matrix of exogenous variables with dimensions \code{t} by \code{m}.}
}
\value{
List containing the dependent variable (Y)
and predictor variable (X) matrices.
Note that the resulting matrices will have \code{t - p} rows.
}
\description{
This function creates the dependent variable (Y)
and predictor variable (X) matrices.
}
\details{
The \code{\link[=YX]{YX()}} function creates the \code{Y} and \code{X} matrices
required for fitting a Vector Autoregressive (VAR) model.
Given the input \code{data} matrix with dimensions \code{t} by \code{k},
where \code{t} is the number of observations and \code{k} is the number of variables,
and the order of the VAR model \code{p} (number of lags),
the function constructs lagged predictor matrix \code{X}
and the dependent variable matrix \code{Y}.

The steps involved in creating the \code{Y} and \code{X} matrices are as follows:
\itemize{
\item Determine the number of observations \code{t} and the number of variables \code{k}
from the input data matrix.
\item Create matrices \code{X} and \code{Y} to store lagged variables
and the dependent variable, respectively.
\item Populate the matrices \code{X} and \code{Y} with the appropriate lagged data.
The predictors matrix \code{X} contains a column of ones
and the lagged values of the dependent variables,
while the dependent variable matrix \code{Y} contains the original values
of the dependent variables.
\item The function returns a list containing the \code{Y} and \code{X} matrices,
which can be used for further analysis and estimation
of the VAR model parameters.
}
}
\seealso{
The \code{\link[=SimVAR]{SimVAR()}} function for simulating time series data
from a VAR model.

Other Simulation of Autoregressive Data Functions: 
\code{\link{SimAR}()},
\code{\link{SimMVNFixed}()},
\code{\link{SimMVN}()},
\code{\link{SimVARExo}()},
\code{\link{SimVARZIPExo}()},
\code{\link{SimVARZIP}()},
\code{\link{SimVAR}()},
\code{\link{SimVariance}()},
\code{\link{YX}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Simulation of Autoregressive Data Functions}
\keyword{simAutoReg}
\keyword{utils}
